+++ crates/zng/src/setup.rs
@@ -218,5 +218,7 @@ pub mod task {
 ///
 /// See [`zng_wgt_setup::page`] for the full API.
 pub mod page {
-    pub use zng_wgt_setup::page::{EulaPage, EulaTxt, InstallDirPage, StatusPage, StatusTaskInfo, StatusTaskInfoFnArgs, WelcomePage};
+    pub use zng_wgt_setup::page::{
+        EulaPage, EulaTxt, FinishAction, FinishPage, InstallDirPage, StatusPage, StatusTaskInfoFnArgs, TaskInfo, WelcomePage,
+    };
 }